Move build_distance_func to pbc_utils
