+++ tobac/feature_detection.py
@@ -30,7 +30,6 @@ from sklearn.neighbors import BallTree
 import iris
 import xarray as xr
 
-from tobac.tracking import build_distance_function
 from tobac.utils import internal as internal_utils
 from tobac.utils import periodic_boundaries as pbc_utils
 from tobac.utils.general import spectral_filtering
@@ -1541,7 +1540,7 @@ def filter_min_distance(
     # Check if we have PBCs.
     if PBC_flag in ["hdim_1", "hdim_2", "both"]:
         # Note that we multiply by dxy to get the distances in spatial coordinates
-        dist_func = build_distance_function(
+        dist_func = pbc_utils.build_distance_function(
             min_h1 * dxy, max_h1 * dxy, min_h2 * dxy, max_h2 * dxy, PBC_flag
         )
         features_tree = BallTree(feature_locations, metric="pyfunc", func=dist_func)
